requirements.txt
langchain==0.3.7
langchain-openai==0.2.9
langchain-community==0.3.7
langsmith==0.1.147
openai==1.54.4
python-dotenv==1.0.1
tiktoken==0.8.0
httpx==0.27.2